+++ src/QuickPlugin.cpp
@@ -102,6 +102,6 @@ const char* QuickPlugin::getDisplayName() const {
 	return _definition.displayName;
 }
 
-const uint8_t* QuickPlugin::getIcon() const {
+const uint32_t* QuickPlugin::getIcon() const {
 	return _definition.icon;
 }
