+++ src/SelectScreen.hpp
@@ -22,7 +22,7 @@ public:
 		Option          options[16];
 		InitialCallback initialCallback;
 		ChangeCallback  changeCallback;
-		const uint8_t*  icon;
+		const uint32_t* icon;
 	};
 
 	SelectScreen(PluginEnvironment& environment, const Definition& definition);
@@ -34,9 +34,9 @@ public:
 	virtual void onEncoderDown(int32_t count) override;
 	virtual void onEncoderUp(int32_t count) override;
 
-	virtual const char*    getName() const override;
-	virtual const char*    getDisplayName() const override;
-	virtual const uint8_t* getIcon() const override;
+	virtual const char*     getName() const override;
+	virtual const char*     getDisplayName() const override;
+	virtual const uint32_t* getIcon() const override;
 
 protected:
 	const Definition& _definition;
